+++ frontend/src/hooks/usePomodoroTimer.tsx
@@ -1,15 +1,12 @@
 import { useState, useEffect, useContext } from "react";
-import {
-  StartPomo,
-  UpdatePomoSecondsLeft,
-} from "../../../wailsjs/go/backend/App";
+import { StartPomo, UpdatePomoSecondsLeft } from "../../wailsjs/go/backend/App";
 import { formatISO } from "date-fns";
-import { useSnackbarWithAction } from "../../hooks/useSnackbarWithAction";
+import { useSnackbarWithAction } from "./useSnackbarWithAction";
 import { useTimer } from "react-timer-hook";
-import { TimerLabel } from "../../data/TimerLabel";
-import { AppContext, AppState } from "../../context/AppContext";
-import { calculateTimeAndLabel } from "../../util/Utils";
-import { HistoryData } from "../../data/HistoryData";
+import { TimerLabel } from "../data/TimerLabel";
+import { AppContext, AppState } from "../context/AppContext";
+import { calculateTimeAndLabel } from "../util/Utils";
+import { HistoryData } from "../data/HistoryData";
 
 function usePomodoroTimer(
   initDateTime: Date,
